+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 
+- New icon.
 - The static Core-class containing the global fractions a single mark is divided into, and methods to convert cents to whole marks, and string representations.
 - Migration utility functions for creating signatures, and updating the signatures of accounts.
 - Guard against excess decimals on the receive screen amount input.
@@ -9,6 +10,7 @@
 ### Changed
 
 - Increased the marks divisibility to 10000 (ten thousand) penni, to avoid problems later.
+- App name to e-MRK.
 
 ## [4.0.0] - 15-03-2026